backend/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
mysql-connector-python
google-api-python-client
google-auth-oauthlib
sendgrid